refining server files, resolving issues

+++ src/server/src/api/users-router.ts
@@ -2,16 +2,15 @@ import { AccessLevel } from '@overseer/common/models/users';
 import { UserConfigurationService } from '@overseer/common/services';
 import { Request, Response, Router } from 'express';
 
-import asyncRequestHandler from '../utilities/async-request-handler';
-import { RouteAuthorizer } from '../utilities/route-authorizers';
+import asyncRequestHandler from './utilities/async-request-handler';
+import { RouteAuthorizer } from './utilities/route-authorizers';
 
-export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationService: UserConfigurationService): Router {
+export function create(routeAuthorizer: RouteAuthorizer, userConfigurationService: UserConfigurationService): Router {
   const router: Router = Router();
-
-  router.all('*', routerAuthorizer.requireAccessLevel(AccessLevel.Administrator));
 
   router.get(
     '/',
+    routeAuthorizer.requireAccessLevel(AccessLevel.Administrator),
     asyncRequestHandler(async function (_request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.getUsers());
     })
@@ -19,6 +18,7 @@ export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationS
 
   router.get(
     '/:userId',
+    routeAuthorizer.requireAccessLevel(AccessLevel.Administrator),
     asyncRequestHandler(async function (request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.getUser(request.params.userId));
     })
@@ -26,6 +26,7 @@ export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationS
 
   router.put(
     '/',
+    routeAuthorizer.requireAccessLevelAfterSetup(AccessLevel.Administrator),
     asyncRequestHandler(async function (request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.createUser(request.body));
     })
@@ -33,6 +34,7 @@ export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationS
 
   router.post(
     '/',
+    routeAuthorizer.requireAccessLevel(AccessLevel.Administrator),
     asyncRequestHandler(async function (request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.updateUser(request.body));
     })
@@ -40,6 +42,7 @@ export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationS
 
   router.delete(
     '/:userId',
+    routeAuthorizer.requireAccessLevel(AccessLevel.Administrator),
     asyncRequestHandler(async function (request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.deleteUser(request.params.userId));
     })
@@ -47,6 +50,7 @@ export function initialize(routerAuthorizer: RouteAuthorizer, userConfigurationS
 
   router.post(
     '/password',
+    routeAuthorizer.requireAccessLevel(AccessLevel.Administrator),
     asyncRequestHandler(async function (request: Request, response: Response): Promise<void> {
       response.json(await userConfigurationService.changePassword(request.body));
     })
